requirements.txt
streamlit>=1.18.1
google-generativeai>=0.1.0
